Add support for Alpha Blending
This diff makes it so the function that combines
the sources always attempts to apply alpha-
blending from the front image into the background
image.

Since it has to make sure that all sources that
could have an alpha channel do so, there were a
few necessary changes to the image loading
process as well.