Add module to compose multiple videos through CSV
Adds videogen, a module that takes a csv description of a video
and composes the multiple elements to generate it

It also supports dealing with multiple videos at once, based
on a directory structure on disk. This meant that all sources
needed to support being reset, which is also added by this diff.

This diff is a bit bigger than it should be, but it should
not interfere with the core functionality of the repository,
so hopefully that's not too bad.