Add support for writing images out of bounds
Up until this point all images needed to respect the boundaries of the
image. This diff adds support to writing out of bounds, and
automatically cropping all information.

There is however an issue on MarginCombinator where x and y values for
the margins seem to be inverted, and I'm not entirely sure why that is
happening, but have tested this. I didn't want to keep this feature
under wraps, so I'm shipping it, but this needs to be investigated.